Add concentrations class
This replaces the old output method involving nested dictionaries with a new method involving a np.array and indexing dictionaries, similar to how reaction_rates was dealt with.  This reduces storage requirements even further.  The order of the concentrations class unfortunately is not fixed.  It may be worth considering finding a way to make it so.