Add HDF5 routines for results + utility functions
This commit replaces the old pickle/dill format with an HDF5-based
results routine.  The advantage of this new format is portability and
hopefully performance when opendeplete is eventually domain decomposed.

A dictionary-like approach was considered in which the top-level group
was the cell ID, then nuclide, then reaction, with polynomials and
values stored inside. This format is incompatable with compression, and
extensively duplicates strings.  So, the format contains two large data
blocks with indexes described by dictionaries that were converted to
groups + attributes.