Add test suite support to opendeplete
This addition adds a mostly complete test suite for opendeplete.  It
includes unit tests for the more easily unit-testable portions of the
code, integral tests for the integrator routines, and a full system
test to ensure that all components work in concert.

Some portions of the code have been rewritten to reduce difficulty in
testing.

In addition, the predictor_c0 coefficients have been fixed.