Replace total_number with flat structure
This replaces the total_number and number_density arrays with arrays of
flat construction.  This saves a tremendous amount of RAM.  This does
however change the random number sequence by changing nuclide order.
This was tested by forcing the same nuclide order in the older version
of the code and re-generating the test results.

Remove duplicate chain nuclides